requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
python-multipart==0.0.20
pypdf==5.4.0
Pillow==11.2.1
python-dotenv==1.1.1
openai==1.76.0
anthropic==0.51.0
